kde/kwin: Add patch to disable finding of QML module
ecm_find_qmlmodule checks if QML modules are installed, and if not it
warns during build time. This is purely for packager knowledge. Not
finding those packages here doesn't affect the runtime of the
kwin_wayland at all.

We are disabling this due to bug in the qmlplugindump which causes it to
not return and get stuck, See issue #28.

Revert this when QTBUG-70460 is solved upstream.