weston: update to 5.0.0
pmaport recreated on top of Alpine's current aport. The --pixman-type
patches were removed, it would have needed to be rewritten to work with
newer Weston source code. It is a Weston specific workaround for a
problem with the downstream kernels, so this is better fixed in the
kernel source code:
https://wiki.postmarketos.org/wiki/Troubleshooting:display#My_screen_is_red.21

The ABGR patch was rebased and submitted for upstreaming here:
https://gitlab.freedesktop.org/wayland/weston/merge_requests/61

Fixes #136.