Revert "linux-*: use upstream gcc6 [skip ci]"
This reverts commit 580fcddb566d453f8fea7b7976e5cb79084ea580
and increases the pkgrels of all affected linux pmaports.

I have compiled *every single kernel* that was modified with this
commit, and it worked. That took 12 hours. So I'm pretty confident that
this is a good commit. Let's roll it out and go back to stability \o/

I'll kick off the binary repo building directly after pushing this, but
it will take some time until all binary packages are available again.

[skip ci]: it wouldn't finish in time.