linux-infocus-flatfish: fix race condition
Add -j1 to compiling the standby code, which is compiled separately
already. This change seems to make the kernel always compile, I've
tried it 6 times, 3 times of that with pmbootstrap's "--no-ccache"
option. It got past an error about 30 seconds into the build, which
happened roughly 2 out of 10 times:

gcc6-armv6-alpine-linux-muslgnueabihf-ld: cannot find standby.o: No such file or directory

I thought, this was related to gcc6 changes, or to changes in abuild,
but both were not the case.

Grant Miller confirmed that this fixed the build, he was able to
compile the kernel ten times in a row with this commit.