temp/qemu: disable block-curl module
Workaround for a build error on aarch64, where the binary package
repository is currently stuck. Compiling it for aarch64 like this:

$ pmbootstrap build --strict --arch=aarch64 qemu

Resulted in:

>>> qemu: Analyzing dependencies...
ERROR: unsatisfiable constraints:
  openssl-dev-1.1.1a-r0:
    conflicts: libressl-dev-2.7.4-r2[pc:libcrypto=1.1.1a]
               libressl-dev-2.7.4-r2[pc:libssl=1.1.1a]
               libressl-dev-2.7.4-r2[pc:openssl=1.1.1a]
    satisfies: curl-dev-7.62.0-r2[openssl-dev]
               libssh2-dev-1.8.0-r4[pc:libcrypto]
               libssh2-dev-1.8.0-r4[pc:libssl]
               spice-dev-0.14.1-r3[pc:openssl]
  libressl-dev-2.7.4-r2:
    conflicts: openssl-dev-1.1.1a-r0[pc:libcrypto=2.7.4]
               openssl-dev-1.1.1a-r0[pc:libssl=2.7.4]
               openssl-dev-1.1.1a-r0[pc:openssl=2.7.4]
    satisfies: world[libressl-dev]
               libssh2-dev-1.8.0-r4[pc:libcrypto]
               libssh2-dev-1.8.0-r4[pc:libssl]
               spice-dev-0.14.1-r3[pc:openssl]
>>> ERROR: qemu: builddeps failed

Note that the only package not mentioned in both "satisfies" outputs is
curl-dev. The real questions are: why is libressl-dev getting pulled in
at all? (Alpine switched back from libressl to openssl, so this should
not happen). And why does this only happen for aarch64, but not for
x86_64 and armhf? But at least this patch unblocks the package builder.