debug-shell: Skip fbsplash properly (!124)
The show_splash function called by the 20-debug-shell.sh script already
skips fbsplash when the device is configured to not use the framebuffer.
But the variable was not defined in the script. This is solved by
including the /etc/deviceinfo file.

Follow-up to !108, where this change got lost in the rebase.