main/maui*: update apps and plasma-angelfish (!136)
This MR updates the app to the latest commit (currently from 2 days
ago). It also updates plasma-angelfish, but I changed the $pkgrel to
follow the format of the other Git packages.

This also re-enables plasma-angelfish and vvave in the default Plasma
Mobile installation, as they were previously disabled due to
qt5-qtwebengine (which they depend on) requiring a soname bump
upstream.