CI: show meaningful error for outdated pmaports (!167)
When we bump the version in pmaports.cfg and the min. required pmaports
version in the pmbootstrap code, then forks of pmaports.git must be
rebased on pmaports.git master. Otherwise "pmbootstrap init" fails, and
because we have it set to quiet (so the output does not distract from
the interesting output if everything goes fine), there was no meaningful
error message yet.

Wrap the "pmbootstrap init" call and display a helpful message that
tells the user to rebase.