Add postmarketos-ui-shelli (!397)
A framebuffer console plus gesture recognition UI option. Tested locally
building and on nexus 5. The power key works fine for me and the freeze
after sitting for about 5 minutes still happens for me but I don't
consider that necessarily a shelli bug, maybe the code needs to feed a
watchdog or something.