main/kernel-scripts: remove (!329)
This package was a workaround for out-of-tree modules (wireguard). We
have dropped the wireguard kernel module package already, and since
kernel-scripts is failing to build now (it would need to be updated),
let's drop it as well.

If somebody wants to bring back *and maintain* the wireguard
out-of-tree kernel module, I've documented how we had done it here:
https://wiki.postmarketos.org/wiki/Out-of-tree_kernel_modules

[ci:skip-build]: building the kernels that have been adjusted to drop
                 kernel-scripts related code would take too long for CI,
                 but I'm making sure that everything still builds before
                 merging.