main/gcc{4,6}, cross/gcc{4,6}*: fix depends
Explicitly depend on mpc1. Our generated gcc aports use the !tracedeps
option, so we need to explicitly set the libraries it depends on.

This has mostly not been an issue, as we are installing our gcc
packages together with Alpine's gcc package, which causes the libraries
to get installed anyway.

Fixes #236.

[ci:skip-build]: takes too long to build.