lg-vee7e: new device (LG Optimus L7 II) (!358)
This is the beginning of a port to the the LG devices with the codename
vee7e, the one i have is the LG-P710 but according to the xda forum
there are some others, so if someone could test those that would be
nice.

I have not done any work besides making the kernel build and checking if
it boots so there still are some rather obvious issues but it boots and
has working USB-networking as well as a partially working screen and
touchscreen.

[ci:skip-build]: already built successfully in CI