temp/mesa: bump pkgrel
Fix up for !359: the pkgrel was not bumped, and because the real mesa
version was stored in _pkgver instead of pkgver (which is set to 9999),
this means that the APKBUILD has exactly the same version as before. So
pmbootstrap will not try to build it, resulting in a missing
mesa-dri-lima package and general confusion.

See also:
https://gitlab.com/postmarketOS/pmaports/merge_requests/359#note_170054969