main/onboard: new aport (!370)
Currently we're using matchbox-keyboard as on screen keyboard. However,
onboard seems to be a better alternative to, it has auto-show, word
suggestions, themes and settings you can choose. So far the best OSK out
there on the market.

There are a few minor issue such as word suggestions doesn't work (not
sure why, but i assume that onboard doesn't know about the program that
was in front of it), sound feedback doesn't work at all due to missing
sound despite it exists in
/usr/share/sounds/freedesktop/stereo/onboard-key-feedback.oga. Other
than that, I haven't found any issue while testing.

Tested on qemu-amd64.