device/pine-dontbeevil: Seperate version for longts (!407)
This adds a seperate dts for the longts version of the devkit (which was
the un-inverted version before) and adds a shortts version which was
inverted before, but didn't exist in pmaports yet because I manually
edited the dts for every release.

This MR abuses the kernel switching feature to switch the dtb instead so
shortts/longts can be selected in the init step.

[ci:skip-build]: build won't finish in time