temp/qemu: remove (!608)
The package was originally added with jemalloc to work around a deadlock
while compiling mesa. I've tested compiling the mesa-git package on
x86_64 for armhf and armv7 and it worked fine. Looking at the original
issue report, the problem only happened with the autotools build system
and not with meson - and as all our mesa or networkmanager aports are
using meson now, I think we can delete this.