kde/calindori: fix kcalendarcore-dev makedepend
Package kcalcore has been renamed to kcalendarcore in Alpine.
pmbootstrap can't resolve kcalcore-dev anymore (not sure if this is a
bug in pmbootstrap, or if the -dev package is not properly replaced in
the upstream APKBUILD). Work around the resulting error by updating the
package to use the new dependency name.