main/postmarketos-mkinitfs: stop forcing full fsck on every boot (!723)
At the moment, a full file system check is forced even if the
file system is marked clean (i.e. it was properly unmounted the
last time it was used). On large file systems with slow I/O
(e.g. SD cards), this can take a significant amount of time.

Removing the -f option of e2fsck allows skipping the full
file system check if it is marked clean, speeding up the boot time.