raspberry-pi: fix sway, phosh startup (!775)
Both sway and phosh require hardware acceleration (DRM) to work. Let's
activate that when we request one of them. Also don't let them fail
during startup on inputless setups (which is common on a raspberry pi).

Phosh is too slow to be unsable, but now, it'll at least be unusable out
of the box ;-)