firmware: Package modem firmware on MSM8916 mainline devices (!800)
MSM8916 has a single "modem" firmware for the audio DSP and modem.
In some early tests, the remote processor boots successfully at least
and the modem shows up in ofono. (Have not done further tests yet...)
However, audio stops working as soon as it is loaded
(because the audio DSP assumes control over the audio hardware).

For now let's just package it for all MSM8916 mainline devices
so it can be installed more easily. It is not used or installed
by default but it can be manually installed for testing.