wileyfox-crackling: use appended DTB instead of QCDT (!907)
The Wileyfox Swift is the only mainline MSM8916 device using QCDT
at the moment. The other devices do not have unique qcom,board-ids
that could be used within QCDT.

So far this has worked fine for the Wileyfox Swift. However,
in preparation to *optionally* load the modem we now need two separate
pairs of DTBs (one with modem and one without modem). This is simple
to implement using appended DTBs, but not possible with QCDT.

So let's stop using QCDT entirely for mainline,
and use appended DTBs for wileyfox-crackling.