temp/mesa-git: list relevant subpackages as depends for -dri- package (!896)
At the moment, every device that wants to make use of mesa-git needs
to depend on all relevant mesa-git subpackages.

We can simplify this by adding these directly as depends for the
-dri- package that most devices will be depending on. That way,
the fact that you need to depend on all relevant subpackages is
mostly hidden away as "implementation detail" in the mesa-git
package, and no special care is required when using mesa-git.