asus-me176c: reduce rootfs size by installing only needed Mesa driver (!910)
The mesa-dri-intel package is deprecated since it was replaced with
mesa-dri-classic and mesa-dri-gallium. Installing mesa-dri-intel
causes both packages to be installed.

The Intel graphics in asus-me176c are not supported by the new
Gallium "iris" driver, therefore asus-me176c can only use the old
i965 driver available in mesa-dri-classic.

Removing mesa-dri-gallium reduces the disk space needed for a minimal
installation on asus-me176c:
  - Before: 329M
  - After:  256M (-73M)