asus-me176c: upgrade mainline kernel to 5.4.20, enable Anbox options (!942)
The patch set for asus-me176c is really small at this point, and there
is no need to make a new tag for each minor kernel update. Usually the
previous patch set will still apply correctly.

Refactor the APKBUILD to build the kernel similar to linux-lts in Alpine,
and use the original kernel tarball together with a small patch file on top.

Also enable Android-related options in the kernel so we can test
Anbox on asus-me176c.