temp/u-boot-pinephone: new aport (!1049)
This adds a package that builds u-boot from the pine64/u-boot repository
which includes patches for enableing less hardware while booting so
there's quicker feedback that the power button has been pressed.

It also has a seperate patch file that modifies the clockspeed for the
memory which is one of the main performance bottlenecks of the A64 SoC.
It's a patch file so it's quick and easy to test out other clock speeds
when building. 600Mhz is stable but it should be able to run up to
624Mhz.