CI: skip all dot folders for now (!1068)
At the moment we assume that all files except in /.* belong to
a package. Now that .shared-patches was moved to device/.shared-patches
that does not work correctly anymore.

This method should really check to which package the files belong
(e.g. walk up directories until it finds an APKBUILD) instead of
assuming they are directly in the same directory as the APKBUILD.

For now just ignore **/.* (i.e. all files in dot folders),
to unblock the MR since it's a critical fix.