maemo: remove (!1126)
The current postmarketOS port does not do Hildon/Maemo Leste justice.
Various features are missing or broken. We can't even upgrade to the newest
version, because of packaging issues that can't easily be resolved. And
nobody is maintaining it currently. So for the people who want to use
Hildon, please use Maemo Leste (https://maemo-leste.github.io/).

If somebody is willing to invest a lot of time for a new postmarketOS
port of Hildon/Maemo Leste, let us know!