samsung-kminilte: disable v4l udev rules (MR 1160)
With the camera driver (VIDEO_EXYNOS_FIMC_IS) enabled, the system
would randomly panic during startup. Even though the camera does
not currently work on this system, the driver is nonetheless enabled
to make the flashlight work.

The panics seem to be triggered by udev's "v4l_id", which queries
capabilities of video4linux devices.

This change adds a udev rules file, to override the behavior of udev's
default rules for v4l.