arrow-db410c: new device (Arrow DragonBoard 410c) (MR 1293)
The DB410c is a SBC with APQ8016 SoC. It runs on mainline without
any patches. This port makes use of the existing linux-postmarketos-qcom-msm8916
kernel package that is used for other mainline MSM8916 devices.

Tested functionality:
  - USB Network, USB Host (e.g. an USB keyboard)
  - Flashing (Fastboot)
  - Display (HDMI)
  - WiFi/BT
  - GPU
  - Audio (HDMI)
  - FDE

Not tested:
  - GPS

Note: The firmware package is called firmware-qcom-db410c (instead of arrow)
because the firmware is provided directly by Qualcomm, not Arrow.