temp/ofono: assume mainline gobi network interface is "rmnet0" (MR 1607)
This allows to connect the modem to the Internet with oFono.
I have verified that this does not break anything if the "rmnet0"
network interface is missing. Plus, all mainline devices currently
covered by the package should also be able to use the new "BAM DMUX"
network driver that is used as network interface to the modem.

(Note: This works differently on newer SoCs, but they also need
       something different in oFono...)