main/linux-postmarketos-allwinner: only RTL8723CS_NEW (MR 1322)
We cannot have two of them at the same time, most compiler would fail here
but for some reason our didn't.

However, it's better to just use one (the new one, or the old one). Enable
CONFIG_RTL8723CS_NEW, which is the new driver that made it to pine64
kernel since 5.5.

Signed-off-by: Danct12 <danct12@disroot.org>

[ci:skip-build]: already built successfully in CI