main/postmarketos-ui-{gnome,phosh}: less revealing epiphany user-agent (MR 1402)
Make it less trivial to fingerprint users of the epiphany browser on
postmarketOS, by replacing the postmarketOS specific user-agent with an
user-agent from a current pixel device.

Old:
	Mozilla/5.0 (postmarketOS device, like iPhone; X11; Linux x86_64) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/13.0 Safari/605.1.15 Epiphany/605.1.15

New:
	Mozilla/5.0 (Linux; Android 10; Pixel) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/83.0.4103.96 Mobile Safari/537.36

Taken from the following page. Note that this seems to be one of the few
cases, where not a full android build ID is included. Which makes it a
bit harder to fingerprint. Not as good as the tor browser one, but with
that one we'd have to pretend that a webkit browser is firefox.

https://developers.whatismybrowser.com/useragents/explore/operating_platform/pixel/2

Related: https://blog.torproject.org/browser-fingerprinting-introduction-and-challenges-ahead
Related: https://developer.chrome.com/multidevice/user-agent
Related: https://webaim.org/blog/user-agent-string-history/