wingtech-wt88047: remove downstream device port (MR 1604)
In preparation for adding mainline device port for Xiaomi Redmi 2
(wt88047), remove the downstream device port. It's still armhf
which makes it clear that it hasn't been tested for a long time.

The mainline port has many more features, also it is aarch64, which
is not compatible with the downstream kernel.