linux-samsung-chagallwifi: switch kernel sources (MR 1672)
With these kernel sources the device boots successfully,
sometimes. Usually, kernel panics and device reboots a couple of
times, but after a few attempts (maybe 1, maybe 7) it succeeds to
boot. The problems seem to be related to FIMC IS, unfortunately it
does not seem to be possible to just disable it in the kernel config,
as that leads to compilation errors.

I have so far just experimented with getting wifi up and running, but
it does not work out of the box even after adding firmware package
with files from TheMuppets, wpa_supplicant fails to connect and dmesg
is filed with lines like:

```
[  112.509235]  [L0: wl_event_handle:  3109] CFG80211-INFO2) wl_escan_handler : Couldn't find P2PIE in probe response/beacon
```

[ci:skip-build]: already built successfully in CI