pmaports.cfg: require pmbootstrap 1.24.0 (MR 1634)
Flashing linux-postmarketos-exynos4 to the i9100 requires pmbootstrap MR
1986.

[ci:skip-vercheck]: linux-samsung-i9300: 3.0.101-r1 instead of -r0 is
not a mistake, the linting error fixes are in a separate commit with
their own pkgrel bump. I decided to leave that in its own commit instead
of squashing together with the other commit (as we usually do), because
the linting fixes are quite a lot and it makes the log more readable.
[ci:skip-build]: already built successfully in CI