main/qtvirtualkeyboard-plasma: fix install dir (MR 1780)
Don't use cmake to copy the files to the target path, because it would
install to the wrong directory:
	/usr/lib/qml/QtQuick/VirtualKeyboard/Styles/Plasma/
instead of:
	/usr/lib/qt5/qml/QtQuick/VirtualKeyboard/Styles/Plasma/

I've spent some time trying to figure out why it does that, but could
not find it right away. This is just a legacy component, so not worth
spending much time on this detail.