device-pine64-pinephone: do not rely on modem-power from kernel (MR 1815)
This adds eg25-manager for managing modem power in userspace instead of
relying on the modem-power stuff in the kernel. The userspace
eg25-manager has proven to be more reliable than using modem-power.

An older setup-modem script is installed for ofono, since eg25-manager
cannot interface with ofono (yet).

The eg25 init script was removed since it only dealt with configuring
the modem-power driver in the kernel

[ci:skip-build]: already built successfully in CI