temp/gtk+3.0: upgrade to 9999_git20210108 (MR 1903)
Fix about dialogs not being adaptive and all other changes by Purism,
which are the reason why we use the fork in the first place. The patches
didn't get applied in the previous version we had packaged in
postmarketOS, because the patches are in debian/patches now.

Remove check-version.py, it's in the source tree now.