bq-picmt: Rename to "bq-paella-downstream" (MR 1912)
The "bq-picmt" device port is actually for the same device as the
"bq-paella" device package in community, just with the downstream
kernel. This is useful occasionally for testing.

However, the name applies that this device package should be used
on the BQ variant of the BQ Aquaris X5 (picmt) and the mainline
package should be used on the Cyanogen variant (paella). Actually
both packages work on both variants since the hardware is the same.

To make that clear, rename "bq-picmt" to "bq-paella-downstream".